fixed the CGFLoat on 64bit issue

+++ PSStackedView/PSStackedViewController.m
@@ -363,7 +363,7 @@ enum {
 }typedef PSSVRoundOption;
 
 - (BOOL)isFloatIndexBetween:(CGFloat)floatIndex {
-    CGFloat intIndex, restIndex;
+    float intIndex, restIndex;
     restIndex = modff(floatIndex, &intIndex);
     BOOL isBetween = fabsf(restIndex - 0.5f) < EPSILON;
     return isBetween;
@@ -380,7 +380,7 @@ enum {
             isValid = contentWidth > [self screenWidth] - self.largeLeftInset;
         }else {
             NSUInteger stackCount = [self.viewControllers count];
-            CGFloat intIndex, restIndex;
+            float intIndex, restIndex;
             restIndex = modff(floatIndex, &intIndex); // split e.g. 1.5 in 1.0 and 0.5
             isValid = stackCount > intIndex && contentWidth > ([self screenWidth] - self.leftInset);
             if (isValid && fabsf(restIndex - 0.5f) < EPSILON) {  // comparing floats -> if so, we have a .5 here
@@ -399,7 +399,7 @@ enum {
 
 - (CGFloat)nearestValidFloatIndex:(CGFloat)floatIndex round:(PSSVRoundOption)roundOption {
     CGFloat roundedFloat;
-    CGFloat intIndex, restIndex;
+    float intIndex, restIndex;
     restIndex = modff(floatIndex, &intIndex);
     
     if (restIndex < 0.5f) {
